Project is finished, however major improvements can be made.
Some inefficient movements occur with manipulator. This is due to avoidance of hitting the ground. Using some external package as MoveIt2 would probably greatly improve efficiency of the proposed solution. Additional thing is reactiveness of manipulator. For some reason it seems not to see gains that were set for closed loop control. Additionally some transition of poses can be done simultaneously (opening gripper and approaching cube), however these are minor details, that can be tackled with easily.